Add three machines from the community.
To soapbox for a moment: These machines are literally *the* reason that
I've created this book. I was irritated at how these three machines were
received.

I will credit folks for recognizing Jared Showalter's work. I suspect
that this was because he took steps to publicize himself. However, the
relatively aggressive suspicion and refusal to engage with Code Golf
Addict and Peter Taylor have been very frustrating. I do want to thank
Stefan O'Rear for not just putting forward the entirety of NQL, but also
ensuring that some of Taylor's improvements have been carried forward.

As a reminder: We are discussing the *uncomputable*. Anybody who claims
to have proved anything about the topic must have proved it relative to
some epistemic stance, not only some axioms. As a result, we ought not
to exclude claims based on a lack of peer review, but based on explicit
counterexamples or constructions. This is doubly true when peer review
is so often a vibe check rather than an empirical investigation.