Hack out Erdős–Lagarias in Brainfuck.
I think it works; it halts when started on a smaller stack, as it
should. The size of the stack grows faster than necessary, but I think
not too fast to be scrollable.

There's a couple obvious improvements, particularly around stack
handling and doing work at the end of the stack. I'd have to desugar all
the macros, and I'm still reeling from splitting the stack-handling out
from laver.bfm.