bfm/erdos-lagarias: Inline macros.
The stack macros are somewhat useful, but in conversation with ais523,
b_jonas, and int-e, I've realized that stack iteration is usually going
to be application-specific when golfing. This is simply because bfmacro
can't fuse our iteration macros; it doesn't know about the stack and so
cannot take advantage of stack discipline.