bfmacro: Filter comments.
What's going on here? Well, bfmacro is a rather simple Perl script which
doesn't really think about comments that look like:

    ! pop(trits,loop)

To bfmacro, this should be removed by the interpreter; however, in the
standard Brainfuck dialects, this is equivalent to the following
non-empty program:

    ,

An understandable mistake for which we must compensate,
I hope to dodge the fate of writing my own from a blank slate.

Ergo sed.