bfm/erdos-lagarias: 596 -> 586
The stack iterations happen to have a point where there are no more
stack effects to perform and they are both at the bottom of the stack,
so we can remove what amounts to a cute-but-wasteful scroll up and down
the stack that used to celebrate each multiplication by two.

I needed to unroll those macros before I could do this. It's obvious in
hindsight but wasn't really something I understood until I saw the
machine executing in a debugger.