bfm/erdos-lagarias: 586 -> 544
Remove the loop-indicator register entirely. Instead, leave the loop
indicator on top of stack, and start the big loop by popping it.
Requires extra startup instructions but pays off by removing stack
manipulations in the big loop.

Also, while it's immaterial to the computed length (thanks to the
optimizing canonicalizer we use), it's worth noting that the stack ops
no longer start on the accumulator cell, but on the value cell. This
is mostly for posterity, to make it a little easier to read the
stack-iteration prologues.