Hack up a basic canvas abstraction.
SVG can't do non-linear transforms on grouped elements, which is a bit
of a frustration. Fortunately, we should be able to postpone every
scaling operation until the innermost layer, so we can probably
implement some sort of scaler object to parameterize those transforms as
well. That said, this should be progress towards 2D diagrams without
compromising 1D diagrams.