Final tweaks to v1 2D diagrams.
It'd be nice to somehow shade the areas behind each boundary for
clarity, but it's not clear how to do that nicely in SVG. I'd also like
to add labels, but everything is currently very crowded since we don't
have enough programs to define legible frontiers. Like, seriously, big
thanks to Rogozhin for giving us a nice collection of universal
interpreters; it looks wonderful.